Work on inventory
