--- Assets/Scripts/Player/InventoryHUD.cs
@@ -1,437 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-using System.Linq;
-using DeCrawl.Utils;
-using DeCrawl.Primitives;
-using DeCrawl.UI;
-
-namespace YLHalf
-{
-    public delegate void ClearCorruption(Vector2Int coordinates);
-
-    public class InventoryHUD : MonoBehaviour
-    {
-        public static event ClearCorruption OnClearCorruption;
-
-        public static InventoryHUD instance { get; private set; }
-
-        Dictionary<string, Lootable> Loots = new Dictionary<string, Lootable>();
-        Dictionary<string, RectTransform> Transforms = new Dictionary<string, RectTransform>();
-
-        [SerializeField]
-        InventorySlotUI slotPrefab;
-
-        [SerializeField]
-        Inventory inventory;
-
-        Dictionary<Vector2Int, InventorySlotUI> Slots = new Dictionary<Vector2Int, InventorySlotUI>();
-
-        private void Awake()
-        {
-            if (instance != null && instance != this)
-            {
-                Destroy(gameObject);
-                return;
-            }
-            if (instance == null) instance = this;
-        }
-
-        private void OnDestroy()
-        {
-            if (instance = this) instance = null;
-        }
-
-        private void OnEnable()
-        {
-            Inventory.OnInventoryChange += Inventory_OnInventoryChange;
-            InventorySlotUI.OnBeginHoverSlot += InventorySlotHUD_OnBeginHoverSlot;
-            InventorySlotUI.OnEndHoverSlot += InventorySlotHUD_OnEndHoverSlot;
-            InventorySlotUI.OnBeginDragLoot += InventorySlotHUD_OnBeginDragLoot;
-            InventorySlotUI.OnDragLoot += InventorySlotHUD_OnDragLoot;
-            InventorySlotUI.OnEndDragLoot += InventorySlotHUD_OnEndDragLoot;
-            InventorySlotUI.OnClickSlot += InventorySlotHUD_OnClickSlot;
-            MasterOfEndings.OnEnding += MasterOfEndings_OnEnding;
-        }
-
-        private void OnDisable()
-        {
-            Inventory.OnInventoryChange -= Inventory_OnInventoryChange;
-            InventorySlotUI.OnBeginHoverSlot -= InventorySlotHUD_OnBeginHoverSlot;
-            InventorySlotUI.OnEndHoverSlot -= InventorySlotHUD_OnEndHoverSlot;
-            InventorySlotUI.OnBeginDragLoot -= InventorySlotHUD_OnBeginDragLoot;
-            InventorySlotUI.OnDragLoot -= InventorySlotHUD_OnDragLoot;
-            InventorySlotUI.OnEndDragLoot -= InventorySlotHUD_OnEndDragLoot;
-            InventorySlotUI.OnClickSlot -= InventorySlotHUD_OnClickSlot;
-            MasterOfEndings.OnEnding -= MasterOfEndings_OnEnding;
-        }
-
-        private void InventorySlotHUD_OnClickSlot(InventorySlotUI slot)
-        {
-            if (
-                inventory.Repairs > 0 &&
-                inventory.RemoveOneCorruption(
-                    slot.Coordinates.XYVector3Int(),
-                    () =>
-                    {
-                        if (inventory.Has(l => l.GetType() == typeof(Repair), out Lootable repair))
-                        {
-                            repair.Loot(LootOwner.None);
-                            return true;
-                        }
-                        return false;
-                    },
-                    out int remaining
-                )
-            )
-            {
-                if (remaining == 0)
-                {
-                    slot.State = InventorySlotUIState.Free;
-                    slot.RomanNumeralCount = 0;
-                }
-                else if (remaining > 0)
-                {
-                    slot.RomanNumeralCount = remaining;
-                }
-            }
-        }
-
-        private void MasterOfEndings_OnEnding(EndingType type, Ending ending)
-        {
-            enabled = false;
-        }
-
-        string dragged;
-        Vector3 mouseDragStart;
-        Vector2Int dragSlotStart;
-        Vector2Int hoverSlot;
-
-        private void InventorySlotHUD_OnBeginDragLoot(string lootId)
-        {
-            dragged = lootId;
-            dragSlotStart = hoverSlot;
-            mouseDragStart = Input.mousePosition;
-            Debug.Log(dragged);
-        }
-
-        private void InventorySlotHUD_OnDragLoot(string lootId)
-        {
-            var offset = Input.mousePosition - mouseDragStart;
-            var rt = Transforms[lootId];
-            var loot = Loots[lootId];
-            SlotPosition(loot.Coordinates, loot.UIShape, rt, offset);
-        }
-
-        private void InventorySlotHUD_OnEndDragLoot(string lootId)
-        {
-            if (dragged != null)
-            {
-                var slotsOffset = hoverSlot - dragSlotStart;
-                ApplyOverLootSlots(dragged, coordinates =>
-                {
-                    var offsetCoordinates = coordinates + slotsOffset;
-                    if (Slots.ContainsKey(offsetCoordinates))
-                    {
-                        Slots[offsetCoordinates].Hover = false;
-                    }
-                });
-            }
-            var rt = Transforms[lootId];
-            var loot = Loots[lootId];
-            dragged = null;
-
-            if (hoverSlot.x < 0)
-            {
-                if (!loot.Loot(LootOwner.Level, PlayerController.instance.Position))
-                {
-                    Debug.LogWarning($"Level never picked up {lootId}");
-                }
-                return;
-            }
-
-            var newSlot = loot.Coordinates.XYVector2Int() + (hoverSlot - dragSlotStart);
-            if (!loot.Loot(LootOwner.Player, newSlot.XYVector3Int()))
-            {
-                SlotPosition(loot.Coordinates, loot.UIShape, rt);
-            }
-        }
-
-        /// <summary>
-        /// Invokes an action over all slots occupied by an inventory loot
-        /// </summary>
-        /// <param name="lootId">The loot in question</param>
-        /// <param name="action">The action to perform</param>
-        void ApplyOverLootSlots(string lootId, System.Action<Vector2Int> action)
-        {
-            var coordinates = Loots[lootId].InventorySlots().ToArray();
-            for (int i = 0; i < coordinates.Length; i++)
-            {
-                action(coordinates[i]);
-            }
-        }
-
-        private void InventorySlotHUD_OnBeginHoverSlot(InventorySlotUI slot)
-        {
-            hoverSlot = slot.Coordinates;
-
-            if (dragged != null)
-            {
-                var slotsOffset = slot.Coordinates - dragSlotStart;
-                ApplyOverLootSlots(dragged, coordinates =>
-                {
-                    var offsetCoordinates = coordinates + slotsOffset;
-                    if (Slots.ContainsKey(offsetCoordinates))
-                    {
-                        Slots[offsetCoordinates].Hover = true;
-                    }
-                });
-            }
-            else if (slot.LootId != null)
-            {
-                ApplyOverLootSlots(slot.LootId, coordinates => { Slots[coordinates].Hover = true; });
-            }
-            else if (slot.State == InventorySlotUIState.Special && inventory.Repairs > 0)
-            {
-                slot.Pulse();
-            }
-        }
-
-        private void InventorySlotHUD_OnEndHoverSlot(InventorySlotUI slot)
-        {
-            hoverSlot = Vector2Int.left;
-            if (dragged != null)
-            {
-                var slotsOffset = slot.Coordinates - dragSlotStart;
-                ApplyOverLootSlots(dragged, coordinates =>
-                {
-                    var offsetCoordinates = coordinates + slotsOffset;
-                    if (Slots.ContainsKey(offsetCoordinates))
-                    {
-                        Slots[offsetCoordinates].Hover = false;
-                    }
-                });
-            }
-            else if (slot.LootId != null)
-            {
-                ApplyOverLootSlots(slot.LootId, coordinates => { Slots[coordinates].Hover = false; });
-            }
-            else if (slot.Pulsing)
-            {
-                slot.StopPulsing();
-            }
-        }
-
-        private void PositionRack(int rackIndex, RectTransform rt)
-        {
-            var rect = GetRackRect(rackIndex);
-
-            rt.anchorMin = rect.min;
-            rt.anchorMax = rect.max;
-            rt.offsetMin = Vector2.zero;
-            rt.offsetMax = Vector2.zero;
-        }
-
-        private void SlotPosition(Vector3Int coordinates, RectInt uiShape, RectTransform rt) => SlotPosition(coordinates, uiShape, rt, Vector3.zero);
-        private void SlotPosition(Vector3Int coordinates, RectInt uiShape, RectTransform rt, Vector3 dragOffset)
-        {
-            int rackIndex = coordinates.y / Inventory.RackHeight;
-
-            var rackRect = GetRackRect(rackIndex);
-            var slotsRect = new Rect(
-                rackRect.xMin + RackPadding.xMin * rackRect.width,
-                rackRect.yMin + RackPadding.yMin * rackRect.height,
-                rackRect.width * RackPadding.width,
-                rackRect.height * RackPadding.height
-            );
-
-            var slotWidth = slotsRect.width / Inventory.RackWidth;
-            var slotHeight = slotsRect.height / Inventory.RackHeight;
-
-            var lootRect = new Rect(
-                slotsRect.xMin + slotWidth * (coordinates.x - uiShape.xMin),
-                slotsRect.yMin + slotHeight * (Inventory.RackHeight - (coordinates.y + 1 + uiShape.height / 2)),
-                slotWidth * uiShape.width,
-                slotHeight * uiShape.height
-            );
-
-            rt.anchorMin = lootRect.min;
-            rt.anchorMax = lootRect.max;
-            rt.offsetMin = dragOffset;
-            rt.offsetMax = dragOffset;
-        }
-
-        static readonly Rect InventoryPadding = new Rect(0.01f, 0.005f, 0.98f, 0.99f);
-
-        private RectTransform CreateChild(string name, Texture2D tex)
-        {
-            var go = new GameObject(name);
-
-            var image = go.AddComponent<Image>();
-            image.preserveAspect = true;
-            image.raycastTarget = false;
-
-            var sprite = Sprite.Create(tex, new Rect(0, 0, tex.width, tex.height), Vector2.one * 0.5f);
-            sprite.name = name;
-            image.sprite = sprite;
-
-            var rt = go.GetComponent<RectTransform>();
-            rt.SetParent(transform);
-            return rt;
-        }
-
-        private Rect GetRackRect(int rackIndex)
-        {
-            var rackHeight = InventoryPadding.height / Inventory.RackLimit;
-            return new Rect(
-                InventoryPadding.xMin,
-                InventoryPadding.yMin + (Inventory.RackLimit - rackIndex - 1) * rackHeight,
-                InventoryPadding.width,
-                rackHeight
-            );
-        }
-
-        private RectTransform CreateRackUI(Lootable loot, InventoryRack rack, Vector3Int placement)
-        {
-            var rackIndex = placement.y;
-
-            var rt = CreateChild($"Rack: {rackIndex}", loot.texture);
-            PositionRack(rackIndex, rt);
-
-            for (
-                int y = rackIndex * Inventory.RackHeight,
-                yMax = (rackIndex + 1) * Inventory.RackHeight,
-                localY = 0;
-                y < yMax;
-                y++, localY++
-             )
-            {
-                for (int x = 0; x < Inventory.RackWidth; x++)
-                {
-                    var coords = new Vector2Int(x, y);
-                    var slot = Instantiate(slotPrefab);
-
-                    slot.gameObject.name = $"Slot {coords}";
-                    slot.transform.SetParent(transform);
-                    slot.Coordinates = coords;
-
-                    if (rack.Occupied[localY, x])
-                    {
-                        slot.State = InventorySlotUIState.Occupied;
-                        slot.RomanNumeralCount = 0;
-                    }
-                    else if (rack.Corruption[localY, x] > 0)
-                    {
-                        slot.State = InventorySlotUIState.Special;
-                        slot.RomanNumeralCount = rack.Corruption[localY, x];
-                    }
-                    else
-                    {
-                        slot.State = InventorySlotUIState.Free;
-                        slot.RomanNumeralCount = 0;
-                    }
-
-                    SlotPosition(new Vector3Int(x, y), new RectInt(0, 0, 1, 1), slot.GetComponent<RectTransform>());
-
-                    Slots.Add(coords, slot);
-                }
-            }
-            return rt;
-        }
-
-        static readonly Rect RackPadding = new Rect(0.05f, 0f, 0.9f, 1f);
-
-        private RectTransform CreateLootUI(Lootable loot, Vector3Int placement)
-        {
-            var rt = CreateChild($"Item: {loot.Id}", loot.texture);
-
-            if (loot.GetType() == typeof(Canister))
-            {
-                var canisterHUD = rt.gameObject.AddComponent<CanisterHUD>();
-                canisterHUD.canister = loot as Canister;
-            }
-            else if (loot.GetType() == typeof(Attack))
-            {
-                var lootHUD = rt.gameObject.AddComponent<AttackInventoryHUD>();
-                lootHUD.attack = loot as Attack;
-            }
-
-            SlotPosition(placement, loot.UIShape, rt);
-            return rt;
-        }
-
-
-        private void ApplyNewSlotState(Lootable loot, Vector3Int placement, InventorySlotUIState state)
-        {
-            var coordinates = loot.InventorySlots(placement).ToArray();
-            for (int i = 0; i < coordinates.Length; i++)
-            {
-                var slot = Slots[coordinates[i]];
-                slot.State = state;
-                    slot.LootId = state == InventorySlotUIState.Free ? null : loot.Id;
-            }
-
-        }
-
-        private void PutInInventory(Lootable loot, Vector3Int placement)
-        {
-            Loots.Add(loot.Id, loot);
-            var rack = loot.GetComponent<InventoryRack>();
-            if (rack != null)
-            {
-                Transforms.Add(loot.Id, CreateRackUI(loot, rack, placement));
-            }
-            else
-            {
-                Transforms.Add(loot.Id, CreateLootUI(loot, placement));
-                ApplyNewSlotState(loot, placement, InventorySlotUIState.Occupied);
-            }
-        }
-
-        private void MoveLoot(Lootable loot, Vector3Int placement)
-        {
-            Debug.Log("Inventory Move");
-            var rectTransform = Transforms[loot.Id];
-            if (loot.GetComponent<InventoryRack>())
-            {
-                PositionRack(placement.y, rectTransform);
-            }
-            else
-            {
-
-                ApplyNewSlotState(loot, loot.Coordinates, InventorySlotUIState.Free);
-                SlotPosition(placement, loot.UIShape, rectTransform);
-                ApplyNewSlotState(loot, placement, InventorySlotUIState.Occupied);
-            }
-        }
-
-        private void DropLoot(Lootable loot)
-        {
-            var lootId = loot.Id;
-            Loots.Remove(loot.Id);
-            var lootTransform = Transforms[lootId];
-            Destroy(lootTransform.gameObject);
-            Transforms.Remove(lootId);
-            ApplyNewSlotState(loot, loot.Coordinates, InventorySlotUIState.Free);
-            Debug.Log("Inventory Drop");
-        }
-
-        private void Inventory_OnInventoryChange(Lootable loot, InventoryEvent inventoryEvent, Vector3Int placement)
-        {
-            switch (inventoryEvent)
-            {
-                case InventoryEvent.PickUp:
-                    PutInInventory(loot, placement);
-                    break;
-                case InventoryEvent.Drop:
-                    DropLoot(loot);
-                    break;
-                case InventoryEvent.Move:
-                    MoveLoot(loot, placement);
-                    break;
-
-            }
-        }
-    }
-}
