+++ Assets/Scripts/Player/InventoryUI.cs.meta
