+++ Assets/deCrawl/Scrips/Primitives/IInventoryBag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0bdd3005d2b8a049aecca801f47ee66
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
