+++ Assets/deCrawl/Scrips/Primitives/InventoryEvent.cs
@@ -1,0 +1,4 @@
+﻿namespace DeCrawl.Primitives
+{
+    public enum InventoryEvent { PickUp, Drop, Move };
+}
