+++ Assets/deCrawl/Scrips/Primitives/InventoryEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6205f0f82cf99174f9434042318035ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
